% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_formats.R
\name{formats}
\alias{formats}
\alias{minimal_template}
\alias{tiny_template}
\alias{abstract_template}
\alias{curriculum_template}
\title{Defines the formats for the outputs of each template}
\usage{
minimal_template(...)

tiny_template(...)

abstract_template(...)

curriculum_template(...)
}
\arguments{
\item{\dots}{Arguments for the *_document functions}
}
\description{
Functions provide the templates for documents
}
\author{
M Mancini
}
